Input: tsc2007 - restore the initial delay before reading values
When the threaded interrupt handler was introduced in commit 377dc55
("Input: tsc2007 - convert to threaded IRQ") the delay before starting
to read was removed. Later the poll_delay parameter has been removed in
commit f261d46 ("Input: tsc2007 - remove unused poll_delay from
platform data"). However this delay is needed to get reliable values
out of the controller, particularly when used with low quality panels.

Note that the delay introduced in starting the soft handler thread is
not taken into account as it didn't seem worth the code complexity.

To avoid changing the behavior of current configs the poll delay is
now 0 per default, it was 1 milliseconds before the threaded IRQ
conversion.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>