Input: sx8634 - Miscellaneous updates and cleanups
Add support for overriding the button sensitivity and threshold values
via a module parameter and sanity check them before programming the IC
with their values. Refactor the SPM read/write start and stop
programming sequences.

Initialize the PWM earlier to prevent corruption of the SPM and use
gpio_request_one() to request and setup the power GPIO as output.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>