HACK: hid: Reset usb device when receiving too many invalid values
Due to a mechanical construction issue, it is possible that the touch creates
events which do not result from a actual touch-event. These events then occur
in large amounts. To prevent this event flooding we reset the device when we
detect this behavior.

This workaround is only applied to one specific device.

Change-Id: Icfb14c842b4d4dacacd89e8049093abde0daaa43
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1071
Reviewed-by: Julian Scheel <julian@jusst.de>